Extract risk drawdown contracts
