+++ quality/refactor_health_report.md
@@ -272,17 +272,22 @@ The current portfolio holdings contract slice moves shared identity/summary cont
 refreshes the governed monetary-float allowlist for the moved holdings fields, and reduces
 `portfolio.py` from 1,464 to 954 lines while preserving OpenAPI contract tests and the 49-line
 longest-function baseline.
+The current risk drawdown contract slice moves drawdown payload contracts into
+`risk_workspace_drawdown.py`, keeps the legacy `app.contracts.risk_workspace` import surface
+intact, refreshes the governed monetary-float allowlist for the moved drawdown-at-risk fields, and
+reduces `risk_workspace.py` from 2,043 to 1,734 lines while preserving risk workspace contract
+tests and the 49-line longest-function baseline.
 
 ## Health Signals
 
 | Area | Current posture | Evidence |
 | --- | --- | --- |
-| Branch hygiene | Healthy | Current holdings contract branch was created from clean `main` after PR #363; final remote/local cleanup remains a post-merge gate |
-| Unit/contract coverage | Healthy | 1,039 unit/contract tests passed in current branch `make check`; focused holdings, portfolio OpenAPI, insight, workflow, and workspace tests passed with 24 tests on this branch |
+| Branch hygiene | Healthy | Current risk drawdown contract branch was created from clean `main` after PR #364; final remote/local cleanup remains a post-merge gate |
+| Unit/contract coverage | Healthy | 1,039 unit/contract tests passed in the latest merged branch `make check`; focused risk drawdown contract and risk workspace service tests passed with 31 tests on this branch |
 | Integration coverage | Healthy | 207 integration tests passed in current branch `make ci` |
 | Total coverage | Healthy | 1,246 coverage tests passed in current branch `make ci`; total coverage is 94.02%, above the 84% floor |
 | Security audit | Governed | `pip-audit` found no known vulnerabilities after the governed `PYSEC-2026-161` exception |
-| Modularity | Improving, incomplete | Longest-function baseline remains 49 lines; recent response/request-context/parser/mapper/contract slices reduce `portfolio.py` to 954 measured lines and leave `portfolio_service.py` at 2,079 measured lines and `performance_workspace_service.py` at 1,704 measured lines; large contracts and several service files remain above 1,000 lines |
+| Modularity | Improving, incomplete | Longest-function baseline remains 49 lines; recent response/request-context/parser/mapper/contract slices reduce `portfolio.py` to 954 measured lines and `risk_workspace.py` to 1,734 measured lines, leaving `portfolio_service.py` at 2,079 measured lines and `performance_workspace_service.py` at 1,704 measured lines; large contracts and several service files remain above 1,000 lines |
 | API governance | Improving, incomplete | 233 OpenAPI paths and 247 operations have summaries, descriptions, operation IDs, tags, and documented 4xx/5xx responses; Spectral remains report-only |
 | Architecture rules | Improving, incomplete | AST boundary tests exist; import-linter is new report-only baseline |
 | Observability | Partial | Health/readiness/metrics/correlation exist; trace/log scoring not enforced |
